t/data/parsing/html-head-1.dat

#data
<html><t:attr name="'a'" value=1><head><link>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <temma attr>
|     name="'a'"
|     value="1"
|   <head>
|     <link>
